+++ ice-20201109/src/Models/ListSearchLibResponseBody/searchLibInfoList.php
@@ -4,10 +4,16 @@
 
 namespace AlibabaCloud\SDK\ICE\V20201109\Models\ListSearchLibResponseBody;
 
+use AlibabaCloud\SDK\ICE\V20201109\Models\ListSearchLibResponseBody\searchLibInfoList\indexInfo;
 use AlibabaCloud\Tea\Model;
 
 class searchLibInfoList extends Model
 {
+    /**
+     * @var indexInfo[]
+     */
+    public $indexInfo;
+
     /**
      * @description The search library.
      *
@@ -30,6 +36,7 @@ class searchLibInfoList extends Model
      */
     public $status;
     protected $_name = [
+        'indexInfo'     => 'IndexInfo',
         'searchLibName' => 'SearchLibName',
         'status'        => 'Status',
     ];
@@ -41,6 +48,15 @@ class searchLibInfoList extends Model
     public function toMap()
     {
         $res = [];
+        if (null !== $this->indexInfo) {
+            $res['IndexInfo'] = [];
+            if (null !== $this->indexInfo && \is_array($this->indexInfo)) {
+                $n = 0;
+                foreach ($this->indexInfo as $item) {
+                    $res['IndexInfo'][$n++] = null !== $item ? $item->toMap() : $item;
+                }
+            }
+        }
         if (null !== $this->searchLibName) {
             $res['SearchLibName'] = $this->searchLibName;
         }
@@ -59,6 +75,15 @@ class searchLibInfoList extends Model
     public static function fromMap($map = [])
     {
         $model = new self();
+        if (isset($map['IndexInfo'])) {
+            if (!empty($map['IndexInfo'])) {
+                $model->indexInfo = [];
+                $n                = 0;
+                foreach ($map['IndexInfo'] as $item) {
+                    $model->indexInfo[$n++] = null !== $item ? indexInfo::fromMap($item) : $item;
+                }
+            }
+        }
         if (isset($map['SearchLibName'])) {
             $model->searchLibName = $map['SearchLibName'];
         }